Make finite difference checker available in package (Closes #647)
