+++ python/sdist/amici/gradient_check.py
@@ -1,0 +1,1 @@
+../../amici/gradient_check.py
